Fix propagation of `ZREP_INC_FLAG` when doing `zrep refresh`
When doing `zrep refresh`, `ZREP_INC_FLAG` setting (on "refreshing" side)
was (prior this commit) ignored, falling back to default setting.

This is because the value `ZREP_INC_FLAG` was not propagated to the other
("sending") side. One way would be to set it in `/etc/default/zrep` on the
"sending" side - but this is both fiddly and sets it for all (that is,
you cannot have one "refreshing" site using `-i` and the other using `-I`)

Better solution is to simply propagate the setting to the other side,
which is what this commit does.